+++ tests/unit/test_helpmsgs.py
@@ -23,24 +23,23 @@
 from __future__ import print_function, absolute_import
 
 import unittest
-
 import shlex
-from subprocess import Popen, PIPE
 from re import findall
+from subprocess import Popen, PIPE
 
 
-# map of all tests to be defined.
+# Map of all tests to be defined.
 # each test is defined as
-#   name
-#   list of
-#     pywbemcli help function to execute
-#     list of text pieces that must be in result
-# TODO Some day we should match the entire test result but lets keep
-#    it simple until this stabilizes.
+#   name,
+#   list of:
+#       list of components of pywbemcli help function to execute
+#       list of text pieces that must be in result
+# TODO ks Mar 17 Some day we should match the entire test result but lets keep
+#    it simple until code stabilizes.
 TESTS_MAP = {  # pylint: disable=invalid-name
     'top': ["", ['--server', '--default_namespace']],
     "class": ["class", ['get', 'invokemethod']],
-    "classget": ["class get", []],
+    "classget": ["class get", ['localonly', 'propertylist']],
     "classenum": ["class enumerate", []],
     "classassoc": ["class associators", []],
     "classref": ["class references", []],
@@ -71,7 +70,15 @@ TESTS_MAP = {  # pylint: disable=invalid-name
     "serverinfo": ["server info", []],
     "serverns": ["server namespaces", []],
     "serverinterop": ["server interop", []],
-    "serverprof": ["server profiles", []]}
+    "serverprof": ["server profiles", []],
+
+    "connection": ["connection", ['show', 'export', 'set']],
+    "connectionshow": ["connection show", []],
+    "connectionexport": ["connection export", []],
+    "connectionset": ["connection set", []],
+    "connectionlist": ["connection list", []],
+    "connectiondelete": ["connection delete", []],
+    "connectioncreate": ["connection create", []], }
 
 
 class ContainerMeta(type):
@@ -79,12 +86,13 @@ class ContainerMeta(type):
 
     def __new__(mcs, name, bases, dict):  # pylint: disable=redefined-builtin
 
-        def gen_test(tname, cmd_str, result_data):
+        def gen_test(test_name, cmd_str, result_data):
             """
-            Defines the test method and returns the method.
+            Defines the test method that we generate for each test
+            and returns the method.
 
-            Test is the method for each test. Each test builds the pywbemcli
-            command executes it and tests the results
+            Each test builds the pywbemcli command executes it and tests the
+            results
             """
             def test(self):  # pylint: disable=missing-docstring
                 command = 'pywbemcli %s --help' % (cmd_str)
@@ -93,25 +101,23 @@ class ContainerMeta(type):
                 out, err = proc.communicate()
                 exitcode = proc.returncode
 
-                self.assertEqual(exitcode, 0, '%s: ExitCond Err, cmd="%s" '
-                                 'exitcode %s' % (tname, command, exitcode))
+                self.assertEqual(exitcode, 0, ('%s: ExitCode Err, cmd="%s" '
+                                               'exitcode %s' %
+                                               (test_name, command, exitcode)))
 
                 # issue 21. The following generates a deprecation warning
-                # during the coverage test.
+                # during the coverage test. Fixed.
                 self.assertEqual(err, "", '%s stderr not empty. returned %s'
-                                 % (tname, err))
+                                 % (test_name, err))
+
                 for item in result_data:
-                    # print('test item:%s\n out:\n%s\n' % (item, out))
                     match_result = findall(item, out)
-                    # print('match_result %s' % match_result)
                     self.assertIsNotNone(match_result,
                                          "Expecting some result")
             return test
 
         for tname, params in TESTS_MAP.iteritems():
             test_name = "test_%s" % tname
-            # print('tname: %s cmd_str: %s, match snippets: %s' %
-            #      (tname, params[0], params[1]))
             dict[test_name] = gen_test(test_name, params[0], params[1])
         return type.__new__(mcs, name, bases, dict)